{"schema":"decodex/commit/1","summary":"split runtime client and completion modules","authority":"manual"}
