+++ apps/decodex/src/agent/tracker_tool_bridge/tools/completion/handlers/closeout.rs
@@ -1,0 +1,80 @@
+use serde_json::Value;
+
+use crate::agent::tracker_tool_bridge::{
+	self, DynamicToolCallResponse, PendingReviewAction, PendingReviewCompletion,
+	ReviewExecutionMode, ReviewHandoffArgs, RunCompletionDisposition, TrackerToolBridge,
+};
+
+impl<'a> TrackerToolBridge<'a> {
+	pub(in crate::agent::tracker_tool_bridge::tools) fn handle_closeout_complete(
+		&self,
+		arguments: Value,
+	) -> DynamicToolCallResponse {
+		let parsed = match serde_json::from_value::<ReviewHandoffArgs>(arguments) {
+			Ok(parsed) => parsed,
+			Err(error) => {
+				return DynamicToolCallResponse::failure(format!(
+					"Invalid `issue.closeout_complete` arguments: {error}"
+				));
+			},
+		};
+
+		if let Err(error) = self.ensure_issue_scope(&parsed.scope) {
+			return DynamicToolCallResponse::failure(error);
+		}
+
+		let Some(review_context) = self.review_context.as_ref() else {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_closeout_complete` is unavailable for this run.",
+			));
+		};
+
+		if review_context.mode != ReviewExecutionMode::Closeout {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_closeout_complete` is unavailable before a retained post-review closeout run starts.",
+			));
+		}
+
+		let pr_url = parsed.pr_url.trim();
+
+		if pr_url.is_empty() {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_closeout_complete` requires a non-empty `pr_url`.",
+			));
+		}
+
+		let summary = tracker_tool_bridge::normalize_summary(&parsed.summary);
+
+		if summary.is_empty() {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_closeout_complete` requires a non-empty `summary`.",
+			));
+		}
+
+		let pull_request = match self.validate_closeout_pr(review_context, pr_url) {
+			Ok(pull_request) => pull_request,
+			Err(error) => return DynamicToolCallResponse::failure(error),
+		};
+
+		if let Err(error) = self.validate_closeout_issue_completed_state() {
+			return DynamicToolCallResponse::failure(error);
+		}
+		if let Err(error) = self.append_review_completion_intent(
+			review_context,
+			RunCompletionDisposition::Closeout,
+			&pull_request,
+			&summary,
+		) {
+			return DynamicToolCallResponse::failure(error);
+		}
+
+		self.pending_review_completion.borrow_mut().replace(PendingReviewCompletion::Closeout(
+			PendingReviewAction { pr_url: pull_request.url.clone(), summary },
+		));
+
+		DynamicToolCallResponse::success(format!(
+			"Recorded retained closeout completion for issue `{}` on merged PR `{}`. Decodex will validate the merged lineage and terminal tracker state before cleaning up the lane.",
+			self.issue.identifier, pull_request.url
+		))
+	}
+}
