+++ apps/decodex/src/agent/tracker_tool_bridge/tools/completion/handlers/repair.rs
@@ -1,0 +1,87 @@
+use serde_json::Value;
+
+use crate::agent::tracker_tool_bridge::{
+	self, DynamicToolCallResponse, ISSUE_REVIEW_REPAIR_COMPLETE_TOOL_NAME, PendingReviewAction,
+	PendingReviewCompletion, ReviewExecutionMode, ReviewHandoffArgs, RunCompletionDisposition,
+	TrackerToolBridge,
+};
+
+impl<'a> TrackerToolBridge<'a> {
+	pub(in crate::agent::tracker_tool_bridge::tools) fn handle_review_repair_complete(
+		&self,
+		arguments: Value,
+	) -> DynamicToolCallResponse {
+		let parsed = match serde_json::from_value::<ReviewHandoffArgs>(arguments) {
+			Ok(parsed) => parsed,
+			Err(error) => {
+				return DynamicToolCallResponse::failure(format!(
+					"Invalid `issue.review_repair_complete` arguments: {error}"
+				));
+			},
+		};
+
+		if let Err(error) = self.ensure_issue_scope(&parsed.scope) {
+			return DynamicToolCallResponse::failure(error);
+		}
+
+		let Some(review_context) = self.review_context.as_ref() else {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_review_repair_complete` is unavailable for this run.",
+			));
+		};
+
+		if review_context.mode != ReviewExecutionMode::Repair {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_review_repair_complete` is unavailable before a retained in-review repair run starts.",
+			));
+		}
+
+		let pr_url = parsed.pr_url.trim();
+
+		if pr_url.is_empty() {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_review_repair_complete` requires a non-empty `pr_url`.",
+			));
+		}
+
+		let summary = tracker_tool_bridge::normalize_summary(&parsed.summary);
+
+		if summary.is_empty() {
+			return DynamicToolCallResponse::failure(String::from(
+				"`issue_review_repair_complete` requires a non-empty `summary`.",
+			));
+		}
+
+		let pull_request = match self.validate_review_action_pr(review_context, pr_url) {
+			Ok(pull_request) => pull_request,
+			Err(error) => return DynamicToolCallResponse::failure(error),
+		};
+
+		if let Err(error) = self.require_clean_review_checkpoint(review_context) {
+			return DynamicToolCallResponse::failure(error);
+		}
+		if let Err(error) = self.clear_review_policy_state_after_completion(
+			review_context,
+			ISSUE_REVIEW_REPAIR_COMPLETE_TOOL_NAME,
+		) {
+			return DynamicToolCallResponse::failure(error);
+		}
+		if let Err(error) = self.append_review_completion_intent(
+			review_context,
+			RunCompletionDisposition::ReviewRepair,
+			&pull_request,
+			&summary,
+		) {
+			return DynamicToolCallResponse::failure(error);
+		}
+
+		self.pending_review_completion.borrow_mut().replace(PendingReviewCompletion::Repair(
+			PendingReviewAction { pr_url: pull_request.url.clone(), summary },
+		));
+
+		DynamicToolCallResponse::success(format!(
+			"Recorded retained review repair completion for issue `{}` on PR `{}`. Decodex will persist the updated review lineage after service validation passes.",
+			self.issue.identifier, pull_request.url
+		))
+	}
+}
